Added order creation for authorized user and improved order types
