+++ src/types/OrderWithAuth.ts
@@ -1,0 +1,8 @@
+export interface IOrderWithAuth {
+  city: string;
+  street?: string;
+  building?: string;
+  apartment?: string;
+  officeNovaPost?: string;
+  comment?: string;
+}
